+++ tests/test_workflow_server.py
@@ -1,71 +1,54 @@
 import json
-import os
-from copy import deepcopy
-from os import path
 from threading import Event
-
+import os
 import walkoff.case.database as case_database
 import walkoff.config.paths
-from walkoff import helpers
-from walkoff.coredb.argument import Argument
 from walkoff.events import WalkoffEvent
-from walkoff.coredb.action import Action
-from walkoff.coredb.branch import Branch
 from walkoff.server import flaskserver as flask_server
 from walkoff.server.returncodes import *
-from tests.util.assertwrappers import orderless_list_compare
 from tests.util.case_db_help import setup_subscriptions_for_action
 from tests.util.servertestcase import ServerTestCase
-import walkoff.coredb.devicedb
+import walkoff.coredb.devicedb as devicedb
 from walkoff.coredb.playbook import Playbook
 from walkoff.coredb.workflow import Workflow
+from walkoff.coredb.action import Action
+from walkoff.coredb.branch import Branch
+from walkoff.coredb.condition import Condition
+from walkoff.coredb.transform import Transform
+from walkoff.coredb.argument import Argument
+from walkoff.coredb.position import Position
+from tests.config import test_workflows_path, test_workflows_path_with_generated
+from uuid import uuid4
 
 
 class TestWorkflowServer(ServerTestCase):
+
     def setUp(self):
-        self.playbooks_added = []
-        self.workflows_added = []
         self.add_playbook_name = 'add_playbook'
         self.change_playbook_name = 'change_playbook'
         self.add_workflow_name = 'add_workflow'
         self.change_workflow_name = 'change_workflow'
-        self.update_playbooks = False
-        self.update_workflows = False
         self.empty_workflow_json = \
             {'actions': [
-                {"app_name": "HelloWorld", "action_name": "helloWorld", "name": "helloworld", "id": -1, "triggers": [],
+                {"app_name": "HelloWorld", "action_name": "helloWorld", "name": "helloworld", "id": str(uuid4()),
+                 "triggers": [],
                  "arguments": []}],
                 'name': self.add_workflow_name,
-                'start': -1,
+                'start': str(uuid4()),
                 'branches': []}
-
-        case_database.initialize()
+        self.verb_lookup = {'get': self.get_with_status_check,
+                            'put': self.put_with_status_check,
+                            'post': self.post_with_status_check,
+                            'delete': self.delete_with_status_check}
 
     def tearDown(self):
-        if self.update_playbooks:
-            playbook = walkoff.coredb.devicedb.device_db.session.query(Playbook).filter_by(
-                name=self.add_playbook_name).first()
-            if playbook:
-                walkoff.coredb.devicedb.device_db.session.delete(playbook)
-            playbook = walkoff.coredb.devicedb.device_db.session.query(Playbook).filter_by(
-                name=self.change_playbook_name).first()
-            if playbook:
-                walkoff.coredb.devicedb.device_db.session.delete(playbook)
+        devicedb.device_db.session.rollback()
+        for class_ in (Playbook, Workflow, Action, Branch, Condition, Transform, Argument, Position):
+            for instance in devicedb.device_db.session.query(class_).all():
+                devicedb.device_db.session.delete(instance)
 
-        if self.update_workflows:
-            workflow = walkoff.coredb.devicedb.device_db.session.query(Workflow).filter_by(
-                name=self.add_workflow_name).first()
-            if workflow:
-                walkoff.coredb.devicedb.device_db.session.delete(workflow)
-            workflow = walkoff.coredb.devicedb.device_db.session.query(Workflow).filter_by(
-                name=self.change_workflow_name).first()
-            if workflow:
-                walkoff.coredb.devicedb.device_db.session.delete(workflow)
-
-        walkoff.coredb.devicedb.device_db.session.commit()
-
+        devicedb.device_db.session.commit()
         case_database.case_db.tear_down()
-        walkoff.coredb.devicedb.device_db.tear_down()
 
     def copy_playbook(self):
         data = {'playbook_name': self.add_playbook_name}
@@ -84,7 +67,7 @@ class TestWorkflowServer(ServerTestCase):
 
     @staticmethod
     def strip_uids(element):
-        element.pop('uid', None)
+        element.pop('id', None)
         for key, value in element.items():
             if isinstance(value, list):
                 for list_element in (list_element_ for list_element_ in value if isinstance(list_element_, dict)):
@@ -92,62 +75,218 @@ class TestWorkflowServer(ServerTestCase):
             elif isinstance(value, dict):
                 for dict_element in (element for element in value.values() if isinstance(element, dict)):
                     TestWorkflowServer.strip_uids(dict_element)
+        return element
 
-    def test_display_all_playbooks(self):
-        num_playbooks = len(walkoff.coredb.devicedb.device_db.session.query(Playbook).all())
+    @staticmethod
+    def load_playbooks(playbooks):
+        paths = []
+        for directory in (test_workflows_path_with_generated, test_workflows_path):
+            paths.extend([os.path.join(directory, filename) for filename in os.listdir(directory)
+                          if filename.endswith('.playbook') and filename.split('.')[0] in playbooks])
+        for path in paths:
+            with open(path, 'r') as playbook_file:
+                playbook = Playbook.create(json.load(playbook_file))
+                devicedb.device_db.session.add(playbook)
+        devicedb.device_db.session.flush()
+
+    def check_invalid_uuid(self, verb, path, element_type, **kwargs):
+        self.verb_lookup[verb](path, headers=self.headers, error='Invalid ID for {}'.format(element_type),
+                               status_code=BAD_REQUEST, **kwargs)
+
+    def check_invalid_id(self, verb, path, element_type, **kwargs):
+        self.verb_lookup[verb](path, error='{} does not exist'.format(element_type.title()), headers=self.headers,
+                                status_code=OBJECT_DNE_ERROR, **kwargs)
+
+    def test_read_all_playbooks(self):
+        playbook_names = ('basicWorkflowTest', 'dataflowTest')
+        self.load_playbooks(playbook_names)
         response = self.get_with_status_check('/api/playbooks', headers=self.headers)
         for playbook in response:
             for workflow in playbook['workflows']:
                 workflow.pop('id')
-
+        self.assertEqual(len(response), len(playbook_names))
         for playbook in response:
-            if playbook['name'] == 'test':
-                self.assertEqual(playbook['workflows'], [{'name': 'helloWorldWorkflow'}])
-            elif playbook['name'] == 'triggerActionWorkflow':
-                self.assertEqual(playbook['workflows'], [{"name": "triggerActionWorkflow"}])
-            elif playbook['name'] == 'pauseWorkflowTest':
-                self.assertEqual(playbook['workflows'], [{"name": "pauseWorkflow"}])
+            self.assertIn(playbook['name'], playbook_names)
 
-        self.assertEqual(len(response), num_playbooks)
+    def standard_load(self):
+        self.load_playbooks(['test', 'dataflowTest'])
+        playbook = devicedb.device_db.session.query(Playbook).filter_by(name='test').first()
+        return playbook
 
-    def test_display_playbook_workflows(self):
-        playbook = walkoff.coredb.devicedb.device_db.session.query(Playbook).filter_by(name='test').first()
+    # All the reads
+
+    def test_read_playbook(self):
+        playbook = self.standard_load()
         response = self.get_with_status_check('/api/playbooks/{}'.format(playbook.id), headers=self.headers)
 
-        workflows = [workflow.read() for workflow in playbook.workflows]
+        expected_workflows = [self.strip_uids(workflow.read()) for workflow in playbook.workflows]
 
         self.assertEqual(response['name'], 'test')
-        self.assertEqual(len(response['workflows']), len(workflows))
-        self.assertListEqual(response['workflows'], workflows)
+        self.assertEqual(len(response['workflows']), len(expected_workflows))
+        self.assertListEqual([self.strip_uids(workflow) for workflow in response['workflows']], expected_workflows)
 
-    def test_display_playbook_workflows_invalid_id(self):
-        self.get_with_status_check('/api/playbooks/0', error='Playbook does not exist.', headers=self.headers,
-                                   status_code=OBJECT_DNE_ERROR)
+    def test_read_playbook_invalid_id(self):
+        self.check_invalid_id('get', '/api/playbooks/{}'.format(uuid4()), 'playbook')
 
-    def test_display_workflow_invalid_id(self):
-        self.get_with_status_check('/api/playbooks/2/workflows/0',
-                                   error='Workflow does not exist',
-                                   headers=self.headers, status_code=OBJECT_DNE_ERROR)
+    def test_read_playbook_invalid_id_format(self):
+        self.check_invalid_uuid('get', '/api/playbooks/0', 'playbook')
 
-    def test_add_playbook(self):
-        expected_playbooks = walkoff.coredb.devicedb.device_db.session.query(Playbook).all()
+    def test_read_playbook_workflows(self):
+        playbook = self.standard_load()
+        response = self.get_with_status_check('/api/playbooks/{}/workflows'.format(playbook.id), headers=self.headers)
+
+        expected_workflows = [self.strip_uids(workflow.read()) for workflow in playbook.workflows]
+
+        self.assertEqual(len(response), len(expected_workflows))
+        self.assertListEqual([self.strip_uids(workflow) for workflow in response], expected_workflows)
+
+    def test_read_playbook_workflows_invalid_id(self):
+        self.check_invalid_id('get', '/api/playbooks/{}/workflows'.format(uuid4()), 'playbook')
+
+    def test_read_playbook_workflows_invalid_id_format(self):
+        self.check_invalid_uuid('get', '/api/playbooks/0/workflows', 'playbook')
+
+    def test_read_workflow(self):
+        playbook = self.standard_load()
+        target_workflow = playbook.workflows[0]
+        response = self.get_with_status_check(
+            '/api/playbooks/{0}/workflows/{1}'.format(playbook.id, target_workflow.id),
+            headers=self.headers)
+        self.assertDictEqual(self.strip_uids(response), self.strip_uids(target_workflow.read()))
+
+    def test_read_workflow_invalid_playbook_id(self):
+        playbook = self.standard_load()
+        target_workflow = playbook.workflows[0]
+        self.check_invalid_id('get', '/api/playbooks/{0}/workflows/{1}'.format(uuid4(), target_workflow.id), 'workflow')
+
+    def test_read_workflow_invalid_workflow_id(self):
+        playbook = self.standard_load()
+        self.check_invalid_id('get', '/api/playbooks/{0}/workflows/{1}'.format(playbook.id, uuid4()), 'workflow')
+
+    def test_read_workflow_invalid_playbook_and_workflow_id(self):
+        self.check_invalid_id('get', '/api/playbooks/{0}/workflows/{1}'.format(uuid4(), uuid4()), 'workflow')
+
+    def test_read_workflow_invalid_playbook_id_format(self):
+        playbook = self.standard_load()
+        target_workflow = playbook.workflows[0]
+        self.check_invalid_uuid('get', '/api/playbooks/1/workflows/{}'.format(target_workflow.id), 'workflow')
+
+    def test_read_workflow_invalid_workflow_id_format(self):
+        playbook = self.standard_load()
+        self.check_invalid_uuid('get', '/api/playbooks/{}/workflows/42'.format(playbook.id), 'workflow')
+
+    def test_read_workflow_invalid_playbook_and_workflow_id_format(self):
+        self.check_invalid_uuid('get', '/api/playbooks/10/workflows/75a', 'workflow')
+
+    # All the deletes
+
+    def test_delete_playbook(self):
+        playbook = self.standard_load()
+        self.delete_with_status_check('/api/playbooks/{}'.format(playbook.id), headers=self.headers)
+
+        self.assertIsNone(
+            devicedb.device_db.session.query(Playbook).filter_by(id=playbook.id).first())
+
+    def test_delete_playbook_invalid_id(self):
+        self.standard_load()
+        previous_num_playbooks = len(devicedb.device_db.session.query(Playbook).all())
+        self.check_invalid_id('delete', '/api/playbooks/{}'.format(uuid4()), 'playbook')
+        self.assertEqual(previous_num_playbooks, len(devicedb.device_db.session.query(Playbook).all()))
+
+    def test_delete_playbook_invalid_id_format(self):
+        self.standard_load()
+        previous_num_playbooks = len(devicedb.device_db.session.query(Playbook).all())
+        self.check_invalid_id('delete', '/api/playbooks/{}'.format(uuid4()), 'playbook')
+        self.assertEqual(previous_num_playbooks, len(devicedb.device_db.session.query(Playbook).all()))
+
+    def test_delete_workflow(self):
+        self.standard_load()
+
+        workflows = [Workflow('wf{}'.format(i), uuid4()) for i in range(2)]
+
+        for workflow in workflows:
+            devicedb.device_db.session.add(workflow)
+
+        target_playbook = Playbook('play1', workflows=workflows)
+        devicedb.device_db.session.add(target_playbook)
+        devicedb.device_db.session.flush()
+        workflow_ids = [workflow.id for workflow in workflows]
+        original_num_playbooks = len(devicedb.device_db.session.query(Playbook).all())
+        self.delete_with_status_check('/api/playbooks/{0}/workflows/{1}'.format(target_playbook.id, workflow_ids[0]),
+                                      headers=self.headers)
+        self.assertEqual(len(list(target_playbook.workflows)), len(workflow_ids) - 1)
+        self.assertNotIn(workflow_ids[0], [workflow.id for workflow in target_playbook.workflows])
+        self.assertEqual(len(devicedb.device_db.session.query(Playbook).all()), original_num_playbooks)
+
+    def test_delete_last_workflow(self):
+        self.standard_load()
+
+        workflow = Workflow('wf', uuid4())
+        devicedb.device_db.session.add(workflow)
+        target_playbook = Playbook('play1', workflows=[workflow])
+        devicedb.device_db.session.add(target_playbook)
+        devicedb.device_db.session.flush()
+        original_num_playbooks = len(devicedb.device_db.session.query(Playbook).all())
+        self.delete_with_status_check('/api/playbooks/{0}/workflows/{1}'.format(target_playbook.id, workflow.id),
+                                      headers=self.headers)
+        self.assertIsNone(devicedb.device_db.session.query(Playbook).filter_by(name='play1').first())
+        self.assertEqual(len(devicedb.device_db.session.query(Playbook).all()), original_num_playbooks - 1)
+
+    def test_delete_workflow_invalid_playbook_id(self):
+        playbook = self.standard_load()
+        original_num_playbooks = len(devicedb.device_db.session.query(Playbook).all())
+        self.check_invalid_id('delete', '/api/playbooks/{0}/workflows/{1}'.format(uuid4(), playbook.workflows[0].id),
+                              'workflow')
+        self.assertEqual(len(devicedb.device_db.session.query(Playbook).all()), original_num_playbooks)
+
+    def test_delete_workflow_invalid_workflow_id(self):
+        playbook = self.standard_load()
+        original_num_playbooks = len(devicedb.device_db.session.query(Playbook).all())
+        self.check_invalid_id('delete', '/api/playbooks/{0}/workflows/{1}'.format(playbook.id, uuid4()), 'workflow')
+        self.assertEqual(len(devicedb.device_db.session.query(Playbook).all()), original_num_playbooks)
+
+    def test_delete_workflow_invalid_playbook_and_workflow_id(self):
+        self.standard_load()
+        original_num_playbooks = len(devicedb.device_db.session.query(Playbook).all())
+        self.check_invalid_id('delete', '/api/playbooks/{0}/workflows/{1}'.format(uuid4(), uuid4()), 'workflow')
+        self.assertEqual(len(devicedb.device_db.session.query(Playbook).all()), original_num_playbooks)
+
+    def test_delete_workflow_invalid_playbook_id_format(self):
+        playbook = self.standard_load()
+        original_num_playbooks = len(devicedb.device_db.session.query(Playbook).all())
+        self.check_invalid_uuid('delete', '/api/playbooks/af/workflows/{}'.format(playbook.workflows[0].id), 'workflow')
+        self.assertEqual(len(devicedb.device_db.session.query(Playbook).all()), original_num_playbooks)
+
+    def test_delete_workflow_invalid_workflow_id_format(self):
+        playbook = self.standard_load()
+        original_num_playbooks = len(devicedb.device_db.session.query(Playbook).all())
+        self.check_invalid_uuid('delete', '/api/playbooks/{}/workflows/37b'.format(playbook.id, uuid4()), 'workflow')
+        self.assertEqual(len(devicedb.device_db.session.query(Playbook).all()), original_num_playbooks)
+
+    def test_delete_workflow_invalid_playbook_and_workflow_id_format(self):
+        self.standard_load()
+        original_num_playbooks = len(devicedb.device_db.session.query(Playbook).all())
+        self.check_invalid_uuid('delete', '/api/playbooks/10/workflows/108', 'workflow')
+        self.assertEqual(len(devicedb.device_db.session.query(Playbook).all()), original_num_playbooks)
+
+    # All the creates
+
+    def test_create_playbook(self):
+        expected_playbooks = devicedb.device_db.session.query(Playbook).all()
         original_length = len(list(expected_playbooks))
         data = {"name": self.add_playbook_name}
         self.update_playbooks = True
         response = self.put_with_status_check('/api/playbooks', headers=self.headers,
                                               status_code=OBJECT_CREATED, data=json.dumps(data),
                                               content_type="application/json")
-
         response.pop('id')
 
         self.assertDictEqual(response, {'name': self.add_playbook_name, 'workflows': []})
-        self.assertEqual(len(list(walkoff.coredb.devicedb.device_db.session.query(Playbook).all())),
+        self.assertEqual(len(list(devicedb.device_db.session.query(Playbook).all())),
                          original_length + 1)
 
-    def test_add_playbook_already_exists(self):
+    def test_create_playbook_already_exists(self):
         data = {"name": self.add_playbook_name}
-        self.playbooks_added.append(self.add_playbook_name)
-        self.update_playbooks = True
         self.put_with_status_check('/api/playbooks',
                                    data=json.dumps(data), headers=self.headers, status_code=OBJECT_CREATED,
                                    content_type="application/json")
@@ -156,110 +295,113 @@ class TestWorkflowServer(ServerTestCase):
                                    data=json.dumps(data), headers=self.headers, status_code=OBJECT_EXISTS_ERROR,
                                    content_type="application/json")
 
-    def test_edit_playbook_name(self):
-        response = self.copy_playbook()
+    def test_create_playbook_bad_id_in_workflow(self):
+        workflow = Workflow('wf1', uuid4())
+        devicedb.device_db.session.add(workflow)
+        devicedb.device_db.session.flush()
+        workflow_json = workflow.read()
+        workflow_json['id'] = 'garbage'
+        data = {'name': self.add_playbook_name, 'workflows': [workflow_json]}
+        self.put_with_status_check('/api/playbooks',
+                                   data=json.dumps(data), headers=self.headers, status_code=BAD_REQUEST,
+                                   content_type="application/json")
 
-        data = {'id': response['id'], 'name': self.change_playbook_name}
+    def test_create_workflow(self):
+        playbook = self.standard_load()
+        initial_workflows_len = len(playbook.workflows)
+
+        response = self.put_with_status_check('/api/playbooks/{}/workflows'.format(playbook.id),
+                                              headers=self.headers, status_code=OBJECT_CREATED,
+                                              data=json.dumps(self.empty_workflow_json),
+                                              content_type="application/json")
+
+        self.empty_workflow_json['id'] = response['id']
+
+        final_workflows = playbook.workflows
+        self.assertEqual(len(final_workflows), initial_workflows_len + 1)
+
+        workflow = next((workflow for workflow in final_workflows if workflow.name == self.add_workflow_name), None)
+        self.assertIsNotNone(workflow)
+
+    def test_create_workflow_invalid_id(self):
+        self.check_invalid_id('put', '/api/playbooks/{}/workflows'.format(uuid4()), 'playbook',
+                              data=json.dumps(self.empty_workflow_json),
+                              content_type="application/json")
+
+    def test_create_workflow_invalid_id_format(self):
+        self.check_invalid_uuid('put', '/api/playbooks/87fgb/workflows', 'playbook',
+                                data=json.dumps(self.empty_workflow_json),
+                                content_type="application/json")
+
+    # All the updates
+
+    def test_update_playbook_name(self):
+        playbook = self.standard_load()
+
+        data = {'id': str(playbook.id), 'name': self.change_playbook_name}
         response = self.post_with_status_check('/api/playbooks',
                                                data=json.dumps(data),
                                                headers=self.headers,
                                                content_type='application/json')
         self.assertEqual(response['name'], self.change_playbook_name)
 
-    def test_edit_playbook_invalid_id(self):
-        expected = walkoff.coredb.devicedb.device_db.session.query(Playbook).all()
-        data = {"id": 0}
-        response = self.app.post('/api/playbooks', headers=self.headers, content_type="application/json",
-                                 data=json.dumps(data))
-        self.assertEqual(response._status_code, 461)
-        self.assertListEqual(walkoff.coredb.devicedb.device_db.session.query(Playbook).all(), expected)
+    def test_update_playbook_invalid_id(self):
+        self.standard_load()
+        expected = {playbook.name for playbook in devicedb.device_db.session.query(Playbook).all()}
+        data = {'id': str(uuid4()), 'name': self.change_playbook_name}
+        self.check_invalid_id('post', '/api/playbooks', 'playbook', content_type="application/json", data=json.dumps(data))
+        self.assertSetEqual({playbook.name for playbook in devicedb.device_db.session.query(Playbook).all()}, expected)
 
-    def test_add_workflow(self):
-        response = self.copy_playbook()
+    def test_update_playbook_invalid_id_format(self):
+        self.standard_load()
+        expected = {playbook.name for playbook in devicedb.device_db.session.query(Playbook).all()}
+        data = {'id': '475', 'name': self.change_playbook_name}
+        self.check_invalid_uuid('post', '/api/playbooks', 'playbook', content_type="application/json", data=json.dumps(data))
+        self.assertSetEqual({playbook.name for playbook in devicedb.device_db.session.query(Playbook).all()}, expected)
 
-        initial_playbooks = walkoff.coredb.devicedb.device_db.session.query(Playbook).all()
-        initial_workflows = next(
-            playbook.workflows for playbook in initial_playbooks if playbook.name == self.add_playbook_name)
-        initial_workflows_len = len(initial_workflows)
-
-        self.update_workflows = True
-        response = self.put_with_status_check('/api/playbooks/{}/workflows'.format(response['id']),
-                                              headers=self.headers, status_code=OBJECT_CREATED,
-                                              data=json.dumps(self.empty_workflow_json),
-                                              content_type="application/json")
-        self.empty_workflow_json['id'] = response['id']
-
-        final_playbooks = walkoff.coredb.devicedb.device_db.session.query(Playbook).all()
-        final_workflows = next(
-            playbook.workflows for playbook in final_playbooks if playbook.name == self.add_playbook_name)
-        self.assertEqual(len(final_workflows), initial_workflows_len + 1)
-
-        workflow = walkoff.coredb.devicedb.device_db.session.query(Workflow).join(Workflow._playbook).filter(
-            Workflow.name == self.add_workflow_name, Playbook.name == self.add_playbook_name).first()
-        self.assertIsNotNone(workflow)
-
-    def test_edit_workflow_name_only(self):
-        response = self.copy_workflow()
-        workflow = walkoff.coredb.devicedb.device_db.session.query(Workflow).filter_by(id=response['id']).first()
+    def test_update_workflow(self):
+        playbook = self.standard_load()
+        workflow = playbook.workflows[0]
         expected_json = workflow.read()
         expected_json['name'] = self.change_workflow_name
-        response = self.post_with_status_check('/api/playbooks/1/workflows',
+        response = self.post_with_status_check('/api/playbooks/{}/workflows'.format(playbook.id),
                                                data=json.dumps(expected_json),
                                                headers=self.headers,
                                                content_type='application/json')
-
         self.assertDictEqual(response, expected_json)
 
         self.assertIsNotNone(
-            walkoff.coredb.devicedb.device_db.session.query(Workflow).filter_by(id=workflow.id).first())
+            devicedb.device_db.session.query(Workflow).filter_by(id=workflow.id).first())
         self.assertIsNone(
-            walkoff.coredb.devicedb.device_db.session.query(Workflow).join(Workflow._playbook).filter(
+            devicedb.device_db.session.query(Workflow).join(Workflow._playbook).filter(
                 Workflow.name == self.add_workflow_name).first())
 
-    def test_edit_workflow_invalid_id(self):
-        playbook = walkoff.coredb.devicedb.device_db.session.query(Playbook).filter_by(name='test').first()
+    def test_update_workflow_invalid_id(self):
+        playbook = self.standard_load()
         workflow = playbook.workflows[0].read()
-        workflow['id'] = 0
-        initial_workflows = walkoff.coredb.devicedb.device_db.session.query(Workflow).all()
-        self.post_with_status_check('/api/playbooks/{}/workflows'.format(playbook.id),
-                                    data=json.dumps(workflow), headers=self.headers, content_type="application/json",
-                                    status_code=OBJECT_DNE_ERROR)
-        final_workflows = walkoff.coredb.devicedb.device_db.session.query(Workflow).all()
+        initial_workflows = devicedb.device_db.session.query(Workflow).all()
+        self.check_invalid_id('post', '/api/playbooks/{}/workflows'.format(uuid4()), 'workflow',
+                              data=json.dumps(workflow), content_type="application/json")
+        final_workflows = devicedb.device_db.session.query(Workflow).all()
         self.assertSetEqual(set(final_workflows), set(initial_workflows))
 
-    def test_delete_playbook(self):
-        response = self.copy_playbook()
-
-        self.delete_with_status_check('/api/playbooks/{}'.format(response['id']), headers=self.headers)
-
-        self.assertIsNone(
-            walkoff.coredb.devicedb.device_db.session.query(Playbook).filter_by(id=response['id']).first())
-
-    def test_delete_playbook_invalid_id(self):
-        response = self.copy_playbook()
-        self.delete_with_status_check('/api/playbooks/0', headers=self.headers, status_code=OBJECT_DNE_ERROR)
-
-        self.assertIsNotNone(
-            walkoff.coredb.devicedb.device_db.session.query(Playbook).filter_by(id=response['id']).first())
-
-    def test_delete_workflow(self):
-        response = self.copy_workflow()
-
-        self.delete_with_status_check('/api/playbooks/1/workflows/{}'.format(response['id']), headers=self.headers)
-        self.assertIsNone(
-            walkoff.coredb.devicedb.device_db.session.query(Workflow).filter_by(id=response['id']).first())
-
-    def test_delete_workflow_invalid_id(self):
-        self.delete_with_status_check('/api/playbooks/1/workflows/0',
-                                      headers=self.headers, status_code=OBJECT_DNE_ERROR)
-
+    def test_update_workflow_invalid_id_format(self):
+        playbook = self.standard_load()
+        workflow = playbook.workflows[0].read()
+        initial_workflows = devicedb.device_db.session.query(Workflow).all()
+        self.check_invalid_uuid('post', '/api/playbooks/874fe/workflows', 'workflow',
+                              data=json.dumps(workflow), content_type="application/json")
+        final_workflows = devicedb.device_db.session.query(Workflow).all()
+        self.assertSetEqual(set(final_workflows), set(initial_workflows))
+    
+    '''
     def test_copy_workflow(self):
         self.update_workflows = True
         response = self.post_with_status_check('/api/playbooks/1/workflows/1/copy',
                                                headers=self.headers, status_code=OBJECT_CREATED,
                                                data=json.dumps({'workflow_name': self.add_workflow_name}),
                                                content_type="application/json")
-        playbook = walkoff.coredb.devicedb.device_db.session.query(Playbook).filter_by(id=1).first()
+        playbook = devicedb.device_db.session.query(Playbook).filter_by(id=1).first()
         self.assertEqual(len(playbook.workflows), 2)
 
         for workflow in playbook.workflows:
@@ -284,12 +426,12 @@ class TestWorkflowServer(ServerTestCase):
         self.post_with_status_check('/api/playbooks/1/workflows/1/copy', data=json.dumps(data),
                                     headers=self.headers, status_code=OBJECT_CREATED, content_type="application/json")
 
-        original_playbook = walkoff.coredb.devicedb.device_db.session.query(Playbook).filter_by(id=1).first()
-        new_playbook = walkoff.coredb.devicedb.device_db.session.query(Playbook).filter_by(id=2).first()
+        original_playbook = devicedb.device_db.session.query(Playbook).filter_by(id=1).first()
+        new_playbook = devicedb.device_db.session.query(Playbook).filter_by(id=2).first()
         for workflow in original_playbook.workflows:
             if workflow.name == 'helloWorldWorkflow':
                 original_workflow = workflow
-        copy_workflow = walkoff.coredb.devicedb.device_db.session.query(Workflow).filter_by(name=self.add_workflow_name).first()
+        copy_workflow = devicedb.device_db.session.query(Workflow).filter_by(name=self.add_workflow_name).first()
 
         self.assertEqual(len(original_playbook.workflows), 1)
         self.assertEqual(len(new_playbook.workflows), 2)
@@ -301,8 +443,8 @@ class TestWorkflowServer(ServerTestCase):
     def test_copy_playbook(self):
         response = self.copy_playbook()
 
-        original_playbook = walkoff.coredb.devicedb.device_db.session.query(Playbook).filter_by(name='test').first()
-        copy_playbook = walkoff.coredb.devicedb.device_db.session.query(Playbook).filter_by(name=self.add_playbook_name).first()
+        original_playbook = devicedb.device_db.session.query(Playbook).filter_by(name='test').first()
+        copy_playbook = devicedb.device_db.session.query(Playbook).filter_by(name=self.add_playbook_name).first()
 
         self.assertIsNotNone(original_playbook)
         self.assertIsNotNone(copy_playbook)
@@ -323,7 +465,7 @@ class TestWorkflowServer(ServerTestCase):
 
     def test_execute_workflow(self):
         sync = Event()
-        workflow = walkoff.coredb.devicedb.device_db.session.query(Workflow).filter_by(id=1).first()
+        workflow = devicedb.device_db.session.query(Workflow).filter_by(id=1).first()
         action_uids = [action.id for action in workflow.actions if action.name == 'start']
         setup_subscriptions_for_action(workflow.id, action_uids)
 
@@ -351,7 +493,7 @@ class TestWorkflowServer(ServerTestCase):
     def test_execute_workflow_pause_resume(self):
         sync = Event()
 
-        workflow = walkoff.coredb.devicedb.device_db.session.query(Workflow).filter_by(name='pauseWorkflow').first()
+        workflow = devicedb.device_db.session.query(Workflow).filter_by(name='pauseWorkflow').first()
 
         action_uids = [action.id for action in workflow.actions if action.name == 'start']
         setup_subscriptions_for_action(workflow.id, action_uids)
@@ -408,7 +550,7 @@ class TestWorkflowServer(ServerTestCase):
     def test_execute_workflow_change_arguments(self):
 
         response = self.copy_workflow()
-        workflow = walkoff.coredb.devicedb.device_db.session.query(Workflow).filter_by(id=response['id']).first()
+        workflow = devicedb.device_db.session.query(Workflow).filter_by(id=response['id']).first()
 
         action_uids = [action.id for action in workflow.actions if action.name == 'start']
         setup_subscriptions_for_action(workflow.id, action_uids)
@@ -435,7 +577,7 @@ class TestWorkflowServer(ServerTestCase):
 
     def test_read_results(self):
 
-        workflow = walkoff.coredb.devicedb.device_db.session.query(Workflow).filter_by(id=1).first()
+        workflow = devicedb.device_db.session.query(Workflow).filter_by(id=1).first()
         workflow.execute('a')
         workflow.execute('b')
         workflow.execute('c')
@@ -444,7 +586,7 @@ class TestWorkflowServer(ServerTestCase):
         self.assertSetEqual(set(response.keys()), {'status', 'id', 'results', 'started_at', 'completed_at', 'name'})
 
     def test_read_all_results(self):
-        workflow = walkoff.coredb.devicedb.device_db.session.query(Workflow).filter_by(id=1).first()
+        workflow = devicedb.device_db.session.query(Workflow).filter_by(id=1).first()
 
         workflow.execute('a')
         workflow.execute('b')
@@ -463,7 +605,7 @@ class TestWorkflowServer(ServerTestCase):
 
     def test_execute_workflow_trigger_action(self):
         sync = Event()
-        workflow = walkoff.coredb.devicedb.device_db.session.query(Workflow).filter_by(id=1).first()
+        workflow = devicedb.device_db.session.query(Workflow).filter_by(id=1).first()
         action_uids = [action.id for action in workflow.actions if action.name == 'start']
         setup_subscriptions_for_action(workflow.id, action_uids)
 
@@ -488,3 +630,4 @@ class TestWorkflowServer(ServerTestCase):
         sync.wait(timeout=10)
         self.assertEqual(result['count'], 1)
         self.assertDictEqual(result['data'], {'status': 'Success', 'result': 'REPEATING: Hello World'})
+    '''