+++ walkoff/coredb/branch.py
@@ -7,16 +7,16 @@ from sqlalchemy.orm import relationship, backref
 from walkoff.coredb import Device_Base
 from walkoff.events import WalkoffEvent
 from walkoff.coredb.executionelement import ExecutionElement
-
+from walkoff.dbtypes import Guid
 logger = logging.getLogger(__name__)
 
 
 @total_ordering
 class Branch(ExecutionElement, Device_Base):
     __tablename__ = 'branch'
-    _workflow_id = Column(Integer, ForeignKey('workflow.id'))
-    source_id = Column(Integer, nullable=False)
-    destination_id = Column(Integer, nullable=False)
+    _workflow_id = Column(Guid(), ForeignKey('workflow.id'))
+    source_id = Column(Guid(), nullable=False)
+    destination_id = Column(Guid(), nullable=False)
     status = Column(String(80))
     conditions = relationship('Condition', backref=backref('_branch'), cascade='all, delete-orphan')
     priority = Column(Integer)